part3/requirements.txt
flask
flask-restx
flask-bcrypt
pymysql
mysql-connector-python
mysqlclient
Werkzeug
flask_sqlalchemy
flask_migrate
flask_jwt_extended